requirements.txt
fastapi
pydantic
loguru
uvicorn
revChatGPT
pyyaml
retry